refactor(providers): split vision.py into per-provider files under vision/
