+++ bridge/providers/vision/__init__.py
@@ -1,0 +1,5 @@
+from bridge.providers.vision.coco_detection import Coco2017Detection
+from bridge.providers.vision.image_folder import ImageFolder
+from bridge.providers.vision.torchvision_cifar10 import TorchvisionCIFAR10
+
+__all__ = ["Coco2017Detection", "ImageFolder", "TorchvisionCIFAR10"]
